Refactor report page component and add skeleton loader
